refactor(construct.awscdk.github-pipeline): rename github-pipeline -> workflow
