+++ packages/construct/awscdk/github-pipeline/src/index.ts
@@ -1,3 +1,3 @@
-export * from './github-pipeline'
+export * from './workflow'
 export * from './s3'
 export * from './checkout'
